Update HuffmanCoder.java
ISSUE: Encoding the file omits the last few bits of bits < 8 (The size of a byte). This results in missing the last byte in a file (i.e the last character in a .txt file). It's not obvious in non txt files but sometimes it can corrupt the file.

ISSUE: It's kind of slow. 10 seconds to encode a ~5 mb file and ~38 seconds to decode it.